backend/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-bcrypt>=1.0.1
flask-session>=0.8.0
flask-cors>=4.0.0
google-generativeai>=0.8.0
faster-whisper>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
